Automatically backed up by Learn

+++ lib/create.sql
@@ -1,5 +1,5 @@
-CREATE TABLE projects (id INTEGER PRIMARY KEY, title TEXT, category TEXT, funding_goal INTEGER, start_date TEXT, end_date TEXT);
+CREATE TABLE projects (id INTEGER PRIMARY KEY, title TEXT, category TEXT, funding_goal FLOAT, start_date TEXT, end_date TEXT);
 
 CREATE TABLE users (id INTEGER PRIMARY KEY, name TEXT, age INTEGER);
 
-CREATE TABLE pledges (id INTEGER PRIMARY KEY, amount INTEGER, project_id INTEGER, user_id INTEGER);
+CREATE TABLE pledges (id INTEGER PRIMARY KEY, amount FLOAT, project_id INTEGER, user_id INTEGER);
